backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
tensorflow==2.20.0
pillow==11.3.0
python-multipart==0.0.20
numpy==1.26.4
h5py==3.14.0
